First draft of insert_tag().
